Move native layout to native plugin
